add complex app bar
