+++ lib/pages/sign_in_page.dart
@@ -41,7 +41,10 @@ class _SignInPageState extends State<SignInPage> {
   /// Includes things like checking the internet connection or making sure that
   /// text fields have valid values
   void _onPressed() async {
-    if (_key.currentState!.validate()) {
+
+    Navigator.pushNamed(context, RouteManager.loggedInUserProfile);
+
+    /*if (_key.currentState!.validate()) {
       if (await ConnectionHelper.hasConnection()) {
         // This should work now, perhaps there is a better
         // way of doing it but this approach seems to work
@@ -70,7 +73,7 @@ class _SignInPageState extends State<SignInPage> {
         context: context,
         data: 'Some fields are invalid',
       );
-    }
+    }*/
   }
 
   @override
