+++ pubspec.yaml
@@ -21,6 +21,7 @@ dependencies:
   provider: ^6.0.1
   settings_ui: ^1.0.0
   shared_preferences: ^2.0.8
+  side_sheet: ^1.0.0
 
 dev_dependencies:
   flutter_test:
